+++ src/swidden.erl
@@ -1,11 +1,12 @@
 -module(swidden).
 
--export([start/1, start/2, stop/1]).
+-export([start/1, start/2, start/3, stop/1]).
+-export([get_port/1]).
 -export([success/0, success/1]).
 -export([failure/1, failure/2]).
 -export([redirect/1]).
 
--export_type([json_object/0]).
+-export_type([json_object/0, start_opt/0, start_opts/0]).
 
 -define(DEFAULT_HEADER_NAME, <<"x-swd-target">>).
 
@@ -13,15 +14,38 @@
 
 -type json_object() :: jsone:json_object().
 
+-type start_opt() :: {port, inet:port_number()}
+                   | {app, atom()}
+                   | {header_name, binary()}
+                   | {services, [binary()]}
+                   | {interceptor, module()}
+                   | {ip, inet:ip_address()}
+                   | {middlewares, [module()]}.
 
+-type start_opts() :: [start_opt()].
+
+-define(DEFAULT_IP_ADDRESS, {0, 0, 0, 0}).
+-define(DEFAULT_PORT,       8000).
+
+
+-spec start(atom()) -> {ok, pid()} | {error, term()} | no_return().
 start(Name) when is_atom(Name) ->
-    start(Name, []).
+    start(Name, ?DEFAULT_PORT, []).
 
 
-start(Name, Opts) ->
+-spec start(atom(), start_opts()) -> {ok, pid()} | {error, term()} | no_return().
+start(Name, Opts) when is_atom(Name), is_list(Opts) ->
+    Port = proplists:get_value(port, Opts, ?DEFAULT_PORT),
+    StartOpts = lists:keydelete(port, 1, Opts),
+    start(Name, Port, StartOpts).
+
+
+-spec start(atom(), inet:port_number(), start_opts()) -> {ok, pid()} | {error, term()} | no_return().
+start(Name, Port, Opts) ->
     %% FIXME(nakai): error/1 で対応しているが本来は {error, Reason} で返すべき
-    ok = swidden_dispatch:start(Name),
-    ok = swidden_json_schema:start(Name),
+    AppName = proplists:get_value(app, Opts, Name),
+    ok = swidden_dispatch:start(AppName),
+    ok = swidden_json_schema:start(AppName),
 
     HeaderName = proplists:get_value(header_name, Opts, ?DEFAULT_HEADER_NAME),
 
@@ -40,8 +64,7 @@ start(Name, Opts) ->
                            {services, Services},
                            {interceptor, Interceptor}]}]}]),
 
-    IpAddress = proplists:get_value(ip, Opts, {0, 0, 0, 0}),
-    Port = proplists:get_value(port, Opts, 8000),
+    IpAddress = proplists:get_value(ip, Opts, ?DEFAULT_IP_ADDRESS),
 
     ProtoOpts = case proplists:get_value(middlewares, Opts, not_found) of
                     not_found ->
@@ -53,13 +76,13 @@ start(Name, Opts) ->
     %% コード的に意味不明
     Env = ProtoOpts#{env => #{dispatch => Dispatch}},
 
-    cowboy:start_clear({?REF, Port}, [{ip, IpAddress}, {port, Port}], Env).
+    cowboy:start_clear({?REF, Name}, [{ip, IpAddress}, {port, Port}], Env).
 
 
--spec stop(inet:port_number()) -> ok.
-stop(Port) ->
+-spec stop(atom()) -> ok.
+stop(Name) ->
     %% TODO(v); ets 周りも削除する
-    ok = cowboy:stop_listener({?REF, Port}).
+    ok = cowboy:stop_listener({?REF, Name}).
 
 
 -spec success() -> ok.
@@ -78,9 +101,17 @@ failure(Type) when is_binary(Type) ->
     {error, Type}.
 
 
+-spec failure(binary(), map()) -> {error, binary(), map()}.
 failure(Type, Reason) when is_binary(Type) andalso is_map(Reason) ->
     {error, Type, Reason}.
 
 
+-spec redirect(binary()) -> {ok, {redirect, binary()}}.
 redirect(Location) when is_binary(Location) ->
     {ok, {redirect, Location}}.
+
+
+%% 0 番 port でリッスンした場合のポート番号
+-spec get_port(atom()) -> inet:port_number().
+get_port(Name) when is_atom(Name) ->
+    ranch:get_port({?REF, Name}).