+++ src/swidden_app.erl
@@ -2,17 +2,14 @@
 
 -behaviour(application).
 
-%% Application callbacks
 -export([start/2, stop/1]).
 
-%% ===================================================================
-%% Application callbacks
-%% ===================================================================
 
-
+-spec start(application:start_type(), any()) -> {ok, pid()} | {error, term()} | ignore.
 start(_StartType, _StartArgs) ->
     swidden_sup:start_link().
 
 
+-spec stop(any()) -> ok.
 stop(_State) ->
     ok.