+++ src/swidden_sup.erl
@@ -2,28 +2,17 @@
 
 -behaviour(supervisor).
 
-%% API
 -export([start_link/0]).
 
-%% Supervisor callbacks
 -export([init/1]).
 
-%% Helper macro for declaring children of supervisor
--define(CHILD(I, Type), {I, {I, start_link, []}, permanent, 5000, Type, [I]}).
 
-%% ===================================================================
-%% API functions
-%% ===================================================================
-
-
+-spec start_link() -> {ok, pid()} | {error, term()} | ignore.
 start_link() ->
     supervisor:start_link({local, ?MODULE}, ?MODULE, []).
 
 
-%% ===================================================================
-%% Supervisor callbacks
-%% ===================================================================
-
-
+-spec init(any()) ->
+          {ok, {{supervisor:strategy(), non_neg_integer(), pos_integer()}, [supervisor:child_spec()]}}.
 init([]) ->
     {ok, {{one_for_one, 5, 10}, []}}.